+++ source/Calamari/Program.cs
@@ -16,6 +16,9 @@ namespace Calamari
         protected int Execute(string[] args)
         {
             ProxyInitializer.InitializeDefaultProxy();
+
+            Environment.SetEnvironmentVariable("Monkey","Dish", EnvironmentVariableTarget.Process);
+
             RegisterCommandAssemblies();
 
             try
